requirements.txt
pandas
pymongo
snowflake-connector-python
python-dotenv